outbound::policy: Split router into smaller modules
